requirements.txt
torch>=2.1.1
numpy>=1.26.2
fire>=0.5.0
wandb>=0.16.2
python-dotenv>=1.0.1
retry>=0.9.2
tqdm>=4.66.1
pandas